Bug fixes
fixed duplicate addition when list is empty
fixed state and city being the same despite different location
fixed deleteall
replaced prgroessbar

+++ FeelsLikeWeather/app/src/main/java/com/example/kylie/feelslikeweather/models/darkskypojos/DarkSkyForecast.java
@@ -19,7 +19,7 @@ public class DarkSkyForecast implements Serializable {
     private String timezone;
     @SerializedName("offset")
     @Expose
-    private Integer offset;
+    private Double offset;
     @SerializedName("currently")
     @Expose
     private Currently currently;
@@ -63,11 +63,11 @@ public class DarkSkyForecast implements Serializable {
         this.timezone = timezone;
     }
 
-    public Integer getOffset() {
+    public Double getOffset() {
         return offset;
     }
 
-    public void setOffset(Integer offset) {
+    public void setOffset(Double offset) {
         this.offset = offset;
     }
